UPDATE(🖊️): set exception when empty asrs, skidPlatform

+++ src/asrs-history/asrs-history.service.ts
@@ -75,7 +75,6 @@ export class AsrsHistoryService {
       .leftJoinAndSelect('ah.Asrs', 'Asrs')
       .leftJoinAndSelect('ah.Awb', 'Awb')
       .leftJoinAndSelect('Awb.Scc', 'Scc') // awb의 Scc를 반환합니다.
-      // .where('ah.inOutType = :inOutType', { inOutType: 'in' }) // inOutType이 'in'인 경우 필터링
       .orderBy('ah.asrs_id')
       .addOrderBy('ah.id', 'DESC')
       .getMany(); // 또는 getMany()를 사용하여 엔터티로 결과를 가져올 수 있습니다.
@@ -84,15 +83,16 @@ export class AsrsHistoryService {
       (asrsHistory) => (asrsHistory.Asrs as Asrs).id,
     );
     const awbIds = asrsState.map((asrsHistory) => (asrsHistory.Awb as Awb).id);
-
-    const deleteResult = await this.asrsHistoryRepository
-      .createQueryBuilder()
-      .delete()
-      .where('Asrs IN (:...asrsIds)', { asrsIds })
-      .andWhere('Awb IN (:...awbIds)', { awbIds })
-      .execute();
-
-    return deleteResult;
+    if (asrsIds && asrsIds.length > 0 && awbIds && awbIds.length > 0) {
+      const deleteResult = await this.asrsHistoryRepository
+        .createQueryBuilder()
+        .delete()
+        .where('Asrs IN (:...asrsIds)', { asrsIds })
+        .andWhere('Awb IN (:...awbIds)', { awbIds })
+        .execute();
+      return deleteResult;
+    }
+    return '창고가 비었습니다.';
   }
 
   async findAll(query: AsrsHistory & BasicQueryParamDto) {
